requirements.txt
streamlit==1.36.0
Pillow==10.4.0
pytesseract==0.3.10
langdetect==1.0.9
iptcinfo3==2.1.4
